@@ -7,29 +7,62 @@ import (
 	"path/filepath"
 	"text/template"
 
-	"github.com/iancoleman/strcase"
 	"google.golang.org/protobuf/compiler/protogen"
 )
 
 var (
-	//go:embed template/dml.pb.go.tmpl
-	dmlTmpl string
+	//go:embed template/dml_package.pb.go.tmpl
+	dmlPackageTmpl string
+
+	//go:embed template/dml_package_mysql.pb.go.tmpl
+	dmlPackageMysqlTmpl string
+
+	//go:embed template/dml_package_pagination.pb.go.tmpl
+	dmlPackagePaginationTmpl string
 )
 
 func GenerateDDLFiles(version string, gen *protogen.Plugin) error {
+	var (
+		goPackageVisited = make(map[string]struct{})
+	)
+
 	for _, sourceFile := range gen.Files {
 		if !sourceFile.Generate {
 			continue
 		}
 
-		dmlFileSuffix := "_dml.pb.go"
-		generatedDMLFile := gen.NewGeneratedFile(sourceFile.GeneratedFilenamePrefix+dmlFileSuffix, sourceFile.GoImportPath)
-		addGoFileHead(version, generatedDMLFile, gen, sourceFile)
-		generatedDML, err := generateDML(sourceFile)
-		if err != nil {
-			return err
+		goPackageName := getGoPackageName(sourceFile.Proto.Options)
+		if _, ok := goPackageVisited[goPackageName]; !ok {
+			generatedFilenamePrefix := filepath.Join(filepath.Dir(sourceFile.GeneratedFilenamePrefix), goPackageName)
+
+			dmlPackageFileSuffix := "_dml.pb.go"
+			generatedDMLPackageFile := gen.NewGeneratedFile(generatedFilenamePrefix+dmlPackageFileSuffix, sourceFile.GoImportPath)
+			addFileHead(version, generatedDMLPackageFile, gen, nil, "//")
+			generatedDMLPackage, err := generateDMLPackage(goPackageName)
+			if err != nil {
+				return err
+			}
+			generatedDMLPackageFile.P(generatedDMLPackage)
+
+			dmlPackageMysqlFileSuffix := "_dml_mysql.pb.go"
+			generatedDMLPackageMysqlFile := gen.NewGeneratedFile(generatedFilenamePrefix+dmlPackageMysqlFileSuffix, sourceFile.GoImportPath)
+			addFileHead(version, generatedDMLPackageMysqlFile, gen, nil, "//")
+			generatedDMLPackageMysql, err := generateDMLPackageMysql(goPackageName)
+			if err != nil {
+				return err
+			}
+			generatedDMLPackageMysqlFile.P(generatedDMLPackageMysql)
+
+			dmlPackagePaginationFileSuffix := "_dml_pagination.pb.go"
+			generatedDMLPackagePaginationFile := gen.NewGeneratedFile(generatedFilenamePrefix+dmlPackagePaginationFileSuffix, sourceFile.GoImportPath)
+			addFileHead(version, generatedDMLPackagePaginationFile, gen, nil, "//")
+			generatedDMLPackagePagination, err := generateDMLPackagePagination(goPackageName)
+			if err != nil {
+				return err
+			}
+			generatedDMLPackagePaginationFile.P(generatedDMLPackagePagination)
 		}
-		generatedDMLFile.P(generatedDML)
+		goPackageVisited[goPackageName] = struct{}{}
 
 		for _, message := range sourceFile.Messages {
 			mi, err := NewMessageInfo(*message)
@@ -37,11 +70,19 @@ func GenerateDDLFiles(version string, gen *protogen.Plugin) error {
 				return err
 			}
 
-			messageName := strcase.ToSnake(string(message.Desc.Name()))
+			ddlFileSuffix, err := mi.DDLFileSuffix()
+			if err != nil {
+				return err
+			}
 
-			sqlFileSuffix := fmt.Sprintf("_ddl_%s.pb.sql", messageName)
-			generatedSQLFile := gen.NewGeneratedFile(sourceFile.GeneratedFilenamePrefix+sqlFileSuffix, sourceFile.GoImportPath)
-			addSQLFileHead(version, generatedSQLFile, gen, sourceFile)
+			generatedDDLFile := gen.NewGeneratedFile(sourceFile.GeneratedFilenamePrefix+ddlFileSuffix, sourceFile.GoImportPath)
+			switch filepath.Ext(ddlFileSuffix) {
+			case ".js":
+				addFileHead(version, generatedDDLFile, gen, sourceFile, "//")
+			default:
+				addFileHead(version, generatedDDLFile, gen, sourceFile, "--")
+			}
+
 			stmts, err := mi.GenerateDDL()
 			if err == ErrNotSupportedDatastore {
 				continue
@@ -49,36 +90,42 @@ func GenerateDDLFiles(version string, gen *protogen.Plugin) error {
 			if err != nil {
 				return err
 			}
-			generatedSQLFile.P(stmts)
+			generatedDDLFile.P(stmts)
 
-			messageDMLFilepath := fmt.Sprintf("%s_dml_%s.pb.go", sourceFile.GeneratedFilenamePrefix, messageName)
-			gereratedMessageDMLFile := gen.NewGeneratedFile(messageDMLFilepath, sourceFile.GoImportPath)
-			addGoFileHead(version, gereratedMessageDMLFile, gen, sourceFile)
-
-			messageDMLFilename := filepath.Base(messageDMLFilepath)
-			messageDMLMockFilepath := fmt.Sprintf("%s_dml_%s_mock.pb.go", sourceFile.GeneratedFilenamePrefix, messageName)
-			messageDMLMockFilename := filepath.Base(messageDMLMockFilepath)
-			d, err := mi.GenerateDML(messageDMLFilename, messageDMLMockFilename)
+			messageDMLFileSuffix, err := mi.DMLFileSuffix()
 			if err != nil {
 				return err
 			}
-			gereratedMessageDMLFile.P(d)
+
+			messageDMLFilepath := sourceFile.GeneratedFilenamePrefix + messageDMLFileSuffix
+			generatedMessageDMLFile := gen.NewGeneratedFile(messageDMLFilepath, sourceFile.GoImportPath)
+			addFileHead(version, generatedMessageDMLFile, gen, sourceFile, "//")
+
+			messageDMLMockFileSuffix, err := mi.DMLMockFileSuffix()
+			if err != nil {
+				return err
+			}
+			messageDMLMockFilepath := sourceFile.GeneratedFilenamePrefix + messageDMLMockFileSuffix
+			d, err := mi.GenerateDML(filepath.Base(messageDMLFilepath), filepath.Base(messageDMLMockFilepath))
+			if err != nil {
+				return err
+			}
+			generatedMessageDMLFile.P(d)
 		}
 	}
 
 	return nil
 }
 
-func generateDML(sourceFile *protogen.File) (string, error) {
-	tmpl, err := template.New("dmlTmpl").Parse(dmlTmpl)
+func generateDMLPackage(goPackageName string) (string, error) {
+	tmpl, err := template.New("dmlPackageTmpl").Parse(dmlPackageTmpl)
 	if err != nil {
 		return "", err
 	}
 
-	pkgName := getGoPackageName(sourceFile.Proto.Options)
 	var b bytes.Buffer
 	if err := tmpl.Execute(&b, &dml{
-		PackageName: pkgName,
+		PackageName: goPackageName,
 	}); err != nil {
 		return "", err
 	}
@@ -86,27 +133,49 @@ func generateDML(sourceFile *protogen.File) (string, error) {
 	return b.String(), nil
 }
 
-func addSQLFileHead(version string, g *protogen.GeneratedFile, gen *protogen.Plugin, sourceFile *protogen.File) {
-	g.P("-- Code generated by protoc-gen-go-ddl. DO NOT EDIT.")
-	g.P("-- versions:")
-	g.P("--  protoc-gen-go-ddl v", version)
-	g.P("--  protoc            ", protocVersion(gen))
-	if sourceFile.Proto.GetOptions().GetDeprecated() {
-		g.P("-- ", sourceFile.Desc.Path(), " is a deprecated file.")
-	} else {
-		g.P("-- source: ", sourceFile.Desc.Path())
+func generateDMLPackageMysql(goPackageName string) (string, error) {
+	tmpl, err := template.New("dmlPackageTmpl").Parse(dmlPackageMysqlTmpl)
+	if err != nil {
+		return "", err
 	}
+
+	var b bytes.Buffer
+	if err := tmpl.Execute(&b, &dml{
+		PackageName: goPackageName,
+	}); err != nil {
+		return "", err
+	}
+
+	return b.String(), nil
 }
 
-func addGoFileHead(version string, g *protogen.GeneratedFile, gen *protogen.Plugin, sourceFile *protogen.File) {
-	g.P("// Code generated by protoc-gen-go-ddl. DO NOT EDIT.")
-	g.P("// versions:")
-	g.P("//  protoc-gen-go-ddl v", version)
-	g.P("//  protoc            ", protocVersion(gen))
-	if sourceFile.Proto.GetOptions().GetDeprecated() {
-		g.P("// ", sourceFile.Desc.Path(), " is a deprecated file.")
-	} else {
-		g.P("// source: ", sourceFile.Desc.Path())
+func generateDMLPackagePagination(goPackageName string) (string, error) {
+	tmpl, err := template.New("dmlPackagePaginationTmpl").Parse(dmlPackagePaginationTmpl)
+	if err != nil {
+		return "", err
+	}
+
+	var b bytes.Buffer
+	if err := tmpl.Execute(&b, &dml{
+		PackageName: goPackageName,
+	}); err != nil {
+		return "", err
+	}
+
+	return b.String(), nil
+}
+
+func addFileHead(version string, g *protogen.GeneratedFile, gen *protogen.Plugin, sourceFile *protogen.File, lineComment string) {
+	g.P(fmt.Sprintf("%s Code generated by protoc-gen-go-ddl. DO NOT EDIT.", lineComment))
+	g.P(fmt.Sprintf("%s versions:", lineComment))
+	g.P(fmt.Sprintf("%s  protoc-gen-go-ddl v%s", lineComment, version))
+	g.P(fmt.Sprintf("%s  protoc            %s", lineComment, protocVersion(gen)))
+	if sourceFile != nil {
+		if sourceFile.Proto.GetOptions().GetDeprecated() {
+			g.P(fmt.Sprintf("%s %s is a deprecated file.", lineComment, sourceFile.Desc.Path()))
+		} else {
+			g.P(fmt.Sprintf("%s source: %s", lineComment, sourceFile.Desc.Path()))
+		}
 	}
 	g.P()
 }
