@@ -4,6 +4,8 @@ import (
 	"bytes"
 	_ "embed"
 	"fmt"
+	"sort"
+	"strconv"
 	"strings"
 	"text/template"
 
@@ -12,18 +14,29 @@ import (
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/descriptorpb"
 
-	protobufv1 "github.com/taehoio/ddl/gen/go/taehoio/ddl/protobuf/v1"
+	protobufv1 "github.com/taehoio/protoc-gen-go-ddl/gen/go/taehoio/ddl/protobuf/v1"
 )
 
 var (
-	//go:embed template/dml_message.pb.go.tmpl
-	dmlMessageTmpl string
+	//go:embed template/dml_message_mysql.pb.go.tmpl
+	dmlMessageMySQLTmpl string
+
+	//go:embed template/dml_message_mongodb.pb.go.tmpl
+	dmlMessageMongoDBTmpl string
+)
+
+var (
+	dmlMessageTmpl = map[protobufv1.DatastoreType]string{
+		protobufv1.DatastoreType_DATASTORE_TYPE_MYSQL:   dmlMessageMySQLTmpl,
+		protobufv1.DatastoreType_DATASTORE_TYPE_MONGODB: dmlMessageMongoDBTmpl,
+	}
 )
 
 type MessageInfo struct {
 	message protogen.Message
 
 	MessageOptions []MessageOption
+	NestedMessages []*MessageInfo
 
 	Name    string
 	VarName string
@@ -57,6 +70,11 @@ func NewMessageInfo(message protogen.Message) (*MessageInfo, error) {
 		return nil, err
 	}
 
+	nestedMessages, err := listNestedMessages(message)
+	if err != nil {
+		return nil, err
+	}
+
 	fields, err := extractFields(message)
 	if err != nil {
 		return nil, err
@@ -77,12 +95,13 @@ func NewMessageInfo(message protogen.Message) (*MessageInfo, error) {
 		return nil, err
 	}
 
-	messageName := string(message.Desc.Name())
+	messageName := message.GoIdent.GoName
 
 	mi := &MessageInfo{
 		message: message,
 
 		MessageOptions: messageOptions,
+		NestedMessages: nestedMessages,
 
 		Name:    messageName,
 		VarName: strcase.ToLowerCamel(messageName),
@@ -93,6 +112,10 @@ func NewMessageInfo(message protogen.Message) (*MessageInfo, error) {
 		KeyFields: keyFields,
 		Indices:   indices,
 		Uniques:   uniques,
+	}
+
+	if datastore := mi.getDatastoreOption(); datastore != protobufv1.DatastoreType_DATASTORE_TYPE_MONGODB && len(mi.NestedMessages) > 0 {
+		return nil, fmt.Errorf("nested message in %s datastore is not supported", datastore.String())
 	}
 
 	return mi, nil
@@ -111,6 +134,21 @@ func listMessageOptions(m protogen.Message) ([]MessageOption, error) {
 	})
 
 	return messageOptions, nil
+}
+
+func listNestedMessages(m protogen.Message) ([]*MessageInfo, error) {
+	var nestedMessages []*MessageInfo
+
+	for _, msg := range m.Messages {
+		nestedMessage, err := NewMessageInfo(*msg)
+		if err != nil {
+			return nil, err
+		}
+
+		nestedMessages = append(nestedMessages, nestedMessage)
+	}
+
+	return nestedMessages, nil
 }
 
 func extractFields(m protogen.Message) ([]Field, error) {
@@ -142,62 +180,88 @@ func extractKeyFields(fields []Field) ([]Field, error) {
 	return keyFields, nil
 }
 
-func extractIndices(fields []Field) ([]Index, error) {
-	indexMap := make(map[string][]Field)
+type indexInfo struct {
+	name  string
+	field Field
+	order int32
+}
 
-	for _, field := range fields {
+func extractIndices(fields []Field) ([]Index, error) {
+	indexMap := make(map[string][]indexInfo)
+
+	for fieldIdx, field := range fields {
 		for _, opt := range field.Options {
 			if opt.Name == string(protobufv1.E_Index.TypeDescriptor().FullName()) {
-				kvPairs := strings.Split(opt.Value, ",")
-				for _, kvPair := range kvPairs {
-					kv := strings.Split(kvPair, "=")
-					k := kv[0]
-					indexName := kv[1]
-					if k == "name" {
-						indexMap[indexName] = append(indexMap[indexName], field)
-					}
+				ii, err := extractIndexInfo(field, int32(fieldIdx), opt.Value)
+				if err != nil {
+					return nil, err
 				}
+
+				indexMap[ii.name] = append(indexMap[ii.name], *ii)
 			}
 		}
 	}
 
 	var indices []Index
-	for k, v := range indexMap {
+	for indexName, indexInfos := range indexMap {
+		sort.Slice(indexInfos, func(i, j int) bool {
+			return indexInfos[i].order < indexInfos[j].order
+		})
+
+		var fields []Field
+		for _, ii := range indexInfos {
+			fields = append(fields, ii.field)
+		}
+
 		indices = append(indices, Index{
-			Name:   k,
-			Fields: v,
+			Name:   indexName,
+			Fields: fields,
 		})
 	}
+
+	sort.Slice(indices, func(i, j int) bool {
+		return strings.Compare(indices[i].Name, indices[j].Name) < 0
+	})
 
 	return indices, nil
 }
 
 func extractUniques(fields []Field) ([]Unique, error) {
-	uniqueMap := make(map[string][]Field)
+	uniqueMap := make(map[string][]indexInfo)
 
-	for _, field := range fields {
+	for fieldIdx, field := range fields {
 		for _, opt := range field.Options {
 			if opt.Name == string(protobufv1.E_Unique.TypeDescriptor().FullName()) {
-				kvPairs := strings.Split(opt.Value, ",")
-				for _, kvPair := range kvPairs {
-					kv := strings.Split(kvPair, "=")
-					k := kv[0]
-					uniqueName := kv[1]
-					if k == "name" {
-						uniqueMap[uniqueName] = append(uniqueMap[uniqueName], field)
-					}
+				ii, err := extractIndexInfo(field, int32(fieldIdx), opt.Value)
+				if err != nil {
+					return nil, err
 				}
+
+				uniqueMap[ii.name] = append(uniqueMap[ii.name], *ii)
 			}
 		}
 	}
 
 	var uniques []Unique
-	for k, v := range uniqueMap {
+	for indexName, indexInfos := range uniqueMap {
+		sort.Slice(indexInfos, func(i, j int) bool {
+			return indexInfos[i].order < indexInfos[j].order
+		})
+
+		var fields []Field
+		for _, ii := range indexInfos {
+			fields = append(fields, ii.field)
+		}
+
 		uniques = append(uniques, Unique{
-			Name:   k,
-			Fields: v,
+			Name:   indexName,
+			Fields: fields,
 		})
 	}
+
+	sort.Slice(uniques, func(i, j int) bool {
+		return strings.Compare(uniques[i].Name, uniques[j].Name) < 0
+	})
 
 	return uniques, nil
 }
@@ -207,11 +271,19 @@ var (
 )
 
 func (mi MessageInfo) GenerateDDL() (string, error) {
-	tableName := mi.SQLName
-
-	if !mi.supportsMySQL() {
+	switch mi.getDatastoreOption() {
+	case protobufv1.DatastoreType_DATASTORE_TYPE_MYSQL:
+		return mi.generateMySQLDDL()
+	case protobufv1.DatastoreType_DATASTORE_TYPE_MONGODB:
+		return mi.generateMongodbDDL()
+	default:
 		return "", ErrNotSupportedDatastore
 	}
+}
+
+// generateMySQLDDL generates ddl for mysql. MessageInfo must support mysql datastore.
+func (mi MessageInfo) generateMySQLDDL() (string, error) {
+	tableName := mi.SQLName
 
 	var ddlFields []string
 	for _, field := range mi.Fields {
@@ -223,9 +295,12 @@ func (mi MessageInfo) GenerateDDL() (string, error) {
 		keys = append(keys, fmt.Sprintf("`%s`", k.SQLName))
 	}
 
-	ddlCreateTable := fmt.Sprintf("\nCREATE TABLE `%s` (\n\t%v,\n\tPRIMARY KEY (%s)\n);", tableName, strings.Join(ddlFields, ",\n\t"), strings.Join(keys, ", "))
-
 	var stmts []string
+
+	ddlDropTable := fmt.Sprintf("DROP TABLE IF EXISTS `%s`;", tableName)
+	stmts = append(stmts, ddlDropTable)
+
+	ddlCreateTable := fmt.Sprintf("CREATE TABLE `%s` (\n\t%v,\n\tPRIMARY KEY (%s)\n);", tableName, strings.Join(ddlFields, ",\n\t"), strings.Join(keys, ", "))
 	stmts = append(stmts, ddlCreateTable)
 
 	for _, index := range mi.Indices {
@@ -234,7 +309,7 @@ func (mi MessageInfo) GenerateDDL() (string, error) {
 			indexFieldNames = append(indexFieldNames, fmt.Sprintf("`%s`", ifn.SQLName))
 		}
 
-		stmts = append(stmts, fmt.Sprintf("\nCREATE INDEX `%s` ON `%s` (%s);", index.Name, tableName, strings.Join(indexFieldNames, ", ")))
+		stmts = append(stmts, fmt.Sprintf("CREATE INDEX `%s` ON `%s` (%s);", index.Name, tableName, strings.Join(indexFieldNames, ", ")))
 	}
 
 	for _, unique := range mi.Uniques {
@@ -243,20 +318,98 @@ func (mi MessageInfo) GenerateDDL() (string, error) {
 			uniqueFieldNames = append(uniqueFieldNames, fmt.Sprintf("`%s`", ufn.SQLName))
 		}
 
-		stmts = append(stmts, fmt.Sprintf("\nCREATE UNIQUE INDEX `%s` ON `%s` (%s);", unique.Name, tableName, strings.Join(uniqueFieldNames, ", ")))
+		stmts = append(stmts, fmt.Sprintf("CREATE UNIQUE INDEX `%s` ON `%s` (%s);", unique.Name, tableName, strings.Join(uniqueFieldNames, ", ")))
+	}
 
+	return strings.Join(stmts, "\n\n"), nil
+}
+
+// generateMongodbDDL generates ddl for mongodb. MessageInfo must support mongodb datastore.
+func (mi MessageInfo) generateMongodbDDL() (string, error) {
+	collectionName := mi.SQLName
+
+	var stmts []string
+	stmts = append(stmts, fmt.Sprintf(`db.createCollection("%s")`, collectionName))
+
+	var keys []string
+	for _, k := range mi.KeyFields {
+		keys = append(keys, fmt.Sprintf(`"%s":1`, k.SQLName))
+	}
+	stmts = append(stmts, fmt.Sprintf(`db.%s.createIndex({%s})`, collectionName, strings.Join(keys, ",")))
+
+	for _, index := range mi.Indices {
+		var indexFieldNames []string
+		for _, field := range index.Fields {
+			indexFieldNames = append(indexFieldNames, fmt.Sprintf(`"%s":1`, field.SQLName))
+		}
+
+		stmts = append(stmts, fmt.Sprintf(`db.%s.createIndex({%s}, {"name":"%s"})`, collectionName, strings.Join(indexFieldNames, ","), index.Name))
+	}
+
+	for _, unique := range mi.Uniques {
+		var uniqFieldNames []string
+		for _, field := range unique.Fields {
+			uniqFieldNames = append(uniqFieldNames, fmt.Sprintf(`"%s":1`, field.SQLName))
+		}
+
+		stmts = append(stmts, fmt.Sprintf(`db.%s.createIndex({%s}, {"name":"%s","unique":true})`, collectionName, strings.Join(uniqFieldNames, ","), unique.Name))
 	}
 
 	return strings.Join(stmts, "\n"), nil
 }
 
-func (mi MessageInfo) supportsMySQL() bool {
+func (mi MessageInfo) supportsDatastore(datastore protobufv1.DatastoreType) bool {
 	for _, opt := range mi.MessageOptions {
-		if opt.Name == string(protobufv1.E_DatastoreType.TypeDescriptor().FullName()) && opt.Value == protobufv1.DatastoreType_DATASTORE_TYPE_MYSQL.String() {
+		if opt.Name == string(protobufv1.E_DatastoreType.TypeDescriptor().FullName()) && opt.Value == datastore.String() {
 			return true
 		}
 	}
 	return false
+}
+
+func (mi MessageInfo) getDatastoreOption() protobufv1.DatastoreType {
+	for _, opt := range mi.MessageOptions {
+		if opt.Name == string(protobufv1.E_DatastoreType.TypeDescriptor().FullName()) {
+			return protobufv1.DatastoreType(protobufv1.DatastoreType_value[opt.Value])
+		}
+	}
+	return protobufv1.DatastoreType_DATASTORE_TYPE_UNSPECIFIED
+}
+
+// DMLFileSuffix returns file suffix for dml file.
+func (mi MessageInfo) DMLFileSuffix() (string, error) {
+	switch mi.getDatastoreOption() {
+	case protobufv1.DatastoreType_DATASTORE_TYPE_MYSQL:
+		return fmt.Sprintf("_dml_%s_mysql.pb.go", strcase.ToSnake(mi.Name)), nil
+	case protobufv1.DatastoreType_DATASTORE_TYPE_MONGODB:
+		return fmt.Sprintf("_dml_%s_mongodb.pb.go", strcase.ToSnake(mi.Name)), nil
+	default:
+		return "", ErrNotSupportedDatastore
+	}
+}
+
+// DMLMockFileSuffix returns file suffix for dml mock file.
+func (mi MessageInfo) DMLMockFileSuffix() (string, error) {
+	switch mi.getDatastoreOption() {
+	case protobufv1.DatastoreType_DATASTORE_TYPE_MYSQL:
+		return fmt.Sprintf("_dml_%s_mysql_mock.pb.go", strcase.ToSnake(mi.Name)), nil
+	case protobufv1.DatastoreType_DATASTORE_TYPE_MONGODB:
+		return fmt.Sprintf("_dml_%s_mongodb_mock.pb.go", strcase.ToSnake(mi.Name)), nil
+	default:
+		return "", ErrNotSupportedDatastore
+	}
+}
+
+// DDLFileSuffix returns the suffix of ddl file.
+func (mi MessageInfo) DDLFileSuffix() (string, error) {
+	switch mi.getDatastoreOption() {
+	case protobufv1.DatastoreType_DATASTORE_TYPE_MYSQL:
+		return fmt.Sprintf("_ddl_%s_mysql.pb.sql", strcase.ToSnake(mi.Name)), nil
+	case protobufv1.DatastoreType_DATASTORE_TYPE_MONGODB:
+		return fmt.Sprintf("_ddl_%s_mongodb.pb.js", strcase.ToSnake(mi.Name)), nil
+	default:
+		return "", ErrNotSupportedDatastore
+	}
 }
 
 type dml struct {
@@ -283,7 +436,13 @@ func getGoPackageName(opts *descriptorpb.FileOptions) string {
 }
 
 func (mi MessageInfo) GenerateDML(goFilename string, mockFilename string) (string, error) {
-	tmpl, err := template.New("dmlMessageTmpl").Parse(dmlMessageTmpl)
+	datastore := mi.getDatastoreOption()
+
+	if _, ok := dmlMessageTmpl[datastore]; !ok {
+		return "", ErrNotSupportedDatastore
+	}
+
+	tmpl, err := template.New("dmlMessageTmpl").Parse(dmlMessageTmpl[datastore])
 	if err != nil {
 		return "", err
 	}
@@ -302,4 +461,29 @@ func (mi MessageInfo) GenerateDML(goFilename string, mockFilename string) (strin
 	}
 
 	return b.String(), nil
+}
+
+func extractIndexInfo(field Field, fieldIdx int32, indexSpec string) (*indexInfo, error) {
+	ii := &indexInfo{
+		field: field,
+		order: fieldIdx,
+	}
+
+	kvPairs := strings.Split(indexSpec, ",")
+	for _, kvPair := range kvPairs {
+		kv := strings.Split(kvPair, "=")
+		k := kv[0]
+		v := kv[1]
+		if k == "name" {
+			ii.name = v
+		} else if k == "order" {
+			order32, err := strconv.ParseInt(v, 10, 32)
+			if err != nil {
+				return nil, err
+			}
+			ii.order = int32(order32)
+		}
+	}
+
+	return ii, nil
 }
